'houses' & 'houses live' reducers & actions separation
